Adding a test
The test uses ZeroIC to access _zero variable. Also the q-rule is
on purpose dialed to have less q-points than nodes, because in
InitialCondition for nodal variables we loop over nodes and assign node
index into _qp. For this case, we have to make sure we are not accessing
the _zero out-of-bounds.